+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Unforgettable Extractor whitepaper link;
+
 ## [1.0.3] - 2025-09-25
 ### Changed
 - Set up Algolia credentials;